[#1200] - added random walk sampling to benchmark class - moved all statistics of sampling eval to statistics package - moved all runner of statistics of sampling eval to statistics runner package - fixed typos - removed string output in CSVDataSinkTest
