+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/statistics/AverageClusteringCoefficient.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.flink.model.impl.operators.sampling.statistics;
+package org.gradoop.flink.model.impl.operators.statistics;
 
 import org.apache.flink.api.java.DataSet;
 import org.gradoop.common.model.impl.pojo.GraphHead;
@@ -23,7 +23,7 @@ import org.gradoop.flink.model.impl.epgm.LogicalGraph;
 import org.gradoop.flink.model.api.operators.UnaryGraphToGraphOperator;
 import org.gradoop.flink.model.impl.operators.aggregation.functions.count.VertexCount;
 import org.gradoop.flink.model.impl.operators.aggregation.functions.sum.SumVertexProperty;
-import org.gradoop.flink.model.impl.operators.sampling.statistics.functions.AddAverageCCValueToGraphHeadMap;
+import org.gradoop.flink.model.impl.operators.statistics.functions.AddAverageCCValueToGraphHeadMap;
 
 /**
  * Calculates the average local clustering coefficient of a graph and writes it to the graph head.
