+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/statistics/AverageOutgoingDegree.java
@@ -13,39 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.flink.model.impl.operators.sampling.statistics;
+package org.gradoop.flink.model.impl.operators.statistics;
 
 import org.apache.flink.api.java.DataSet;
 import org.gradoop.common.model.impl.pojo.GraphHead;
 import org.gradoop.flink.model.impl.epgm.LogicalGraph;
 import org.gradoop.flink.model.api.operators.UnaryGraphToGraphOperator;
 import org.gradoop.flink.model.impl.operators.aggregation.functions.count.VertexCount;
-import org.gradoop.flink.model.impl.operators.sampling.statistics.functions.AddSumDegreesToGraphHeadCrossFunction;
-import org.gradoop.flink.model.impl.operators.sampling.statistics.functions.CalculateAverageDegree;
-import org.gradoop.flink.model.impl.operators.statistics.IncomingVertexDegrees;
+import org.gradoop.flink.model.impl.operators.sampling.common.SamplingEvaluationConstants;
+import org.gradoop.flink.model.impl.operators.statistics.functions.AddSumDegreesToGraphHeadCrossFunction;
+import org.gradoop.flink.model.impl.operators.statistics.functions.CalculateAverageDegree;
 
 /**
- * Calculates the average incoming degree of a graph and writes it to the graph head.
- * Uses: {@code ceiling( sum(vertex incoming degrees) / |vertices| )}
+ * Calculates the average outgoing degree of a graph and writes it to the graph head.
+ * Uses: {@code ceiling( sum(vertex outgoing degrees) / |vertices| )}
  */
-public class AverageIncomingDegree implements UnaryGraphToGraphOperator {
+public class AverageOutgoingDegree implements UnaryGraphToGraphOperator {
 
   /**
-   * Calculates the average incoming degree of the input graph and writes it to the graph head.
+   * Calculates the average outgoing degree of the input graph and writes it to the graph head.
    *
    * @param graph the input graph
-   * @return LogicalGraph with the average incoming degree value written to the graph head.
+   * @return LogicalGraph with the average outgoing degree value written to the graph head.
    */
   @Override
   public LogicalGraph execute(LogicalGraph graph) {
     graph = graph.aggregate(new VertexCount());
-    DataSet<GraphHead> newGraphHead = new IncomingVertexDegrees().execute(graph)
+    DataSet<GraphHead> newGraphHead = new OutgoingVertexDegrees().execute(graph)
       .sum(1)
       .crossWithTiny(graph.getGraphHead().first(1))
       .with(new AddSumDegreesToGraphHeadCrossFunction(
         SamplingEvaluationConstants.PROPERTY_KEY_SUM_DEGREES))
       .map(new CalculateAverageDegree(
-        SamplingEvaluationConstants.PROPERTY_KEY_AVERAGE_INCOMING_DEGREE));
+        SamplingEvaluationConstants.PROPERTY_KEY_AVERAGE_OUTGOING_DEGREE));
 
     return graph.getConfig().getLogicalGraphFactory()
       .fromDataSets(newGraphHead, graph.getVertices(), graph.getEdges());
