+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/statistics/DegreeCentrality.java
@@ -13,14 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.flink.model.impl.operators.sampling.statistics;
+package org.gradoop.flink.model.impl.operators.statistics;
 
 import org.apache.flink.api.java.DataSet;
 import org.gradoop.common.model.impl.id.GradoopId;
 import org.gradoop.flink.model.api.operators.UnaryGraphToValueOperator;
 import org.gradoop.flink.model.impl.epgm.LogicalGraph;
-import org.gradoop.flink.model.impl.operators.statistics.VertexCount;
-import org.gradoop.flink.model.impl.operators.statistics.VertexDegrees;
 import org.gradoop.flink.model.impl.operators.statistics.functions.CalculateDegreeCentrality;
 import org.gradoop.flink.model.impl.operators.statistics.functions.DegreeDistanceFunction;
 import org.gradoop.flink.model.impl.tuples.WithCount;
