+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/statistics/DegreeCentralityBase.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.flink.model.impl.operators.sampling.statistics;
+package org.gradoop.flink.model.impl.operators.statistics;
 
 /**
  * Base class with constants for DegreeCentrality
