+++ gradoop-flink/src/test/java/org/gradoop/flink/model/impl/operators/sampling/statistics/DegreeCentralityTest.java
@@ -18,6 +18,7 @@ package org.gradoop.flink.model.impl.operators.sampling.statistics;
 import org.apache.flink.api.java.DataSet;
 import org.gradoop.flink.model.GradoopFlinkTestBase;
 import org.gradoop.flink.model.impl.epgm.LogicalGraph;
+import org.gradoop.flink.model.impl.operators.statistics.DegreeCentrality;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
